Refactor vision into a subsystem
